+++ packages/clients/src/utils/uploadFile.ts
@@ -20,6 +20,42 @@ const getDotExt = (file: File) => {
   }
 };
 
+/*
+<?xml version="1.0" encoding="UTF-8"?>
+<Error>
+    <Code>EntityTooSmall</Code>
+    <Message>Your proposed upload is smaller than the minimum allowed object size.</Message>
+    <BucketName>temp</BucketName>
+    <Resource>/temp</Resource>
+    <RequestId>15ABB241607163F9</RequestId>
+    <HostId>7f1064dc-73b1-4736-be6a-5cadd403509a</HostId>
+</Error>
+ */
+
+export class UploadFileError extends Error {
+  public status: number;
+  public code?: string;
+
+  constructor(xhr: XMLHttpRequest) {
+    let message = `Failed to upload file: ${xhr.responseText}`;
+    let code: string | undefined;
+    if (xhr.responseXML) {
+      const msg = xhr.responseXML.getElementsByTagName('Message').item(0);
+      const cd = xhr.responseXML.getElementsByTagName('Code').item(0);
+      if (msg && msg.textContent) {
+        message = msg.textContent;
+      }
+      if (cd && cd.textContent) {
+        code = cd.textContent;
+      }
+    }
+    super(message);
+    this.status = xhr.status;
+    this.name = 'UploadFileError';
+    this.code = code;
+  }
+}
+
 export const uploadFile = (file: File, link: UploadLink): Promise<string> =>
   new Promise((resolve, reject) => {
     const formData = new FormData();
@@ -41,13 +77,7 @@ export const uploadFile = (file: File, link: UploadLink): Promise<string> =>
         const resultingFilename = location!.split('/').pop()!;
         resolve(resultingFilename);
       } else {
-        reject(
-          new Error(
-            `Failed to upload file. Status code: ${xhr.status}. Message: ${
-              xhr.response
-            }`,
-          ),
-        );
+        reject(new UploadFileError(xhr));
       }
     };
     xhr.onerror = reject;