+++ services/web/src/components/named-node-finder/index.ts
@@ -1,1 +1,2 @@
 export * from './NamedNodeFinder';
+export * from './types';