+++ services/web/src/features/media/list/Dropzone.tsx
@@ -1,4 +1,10 @@
-import { MediaKind } from '@whitewater-guide/commons';
+import {
+  MAX_FILE_SIZE,
+  MAX_FILE_SIZE_STRING,
+  MediaKind,
+  MIN_FILE_SIZE,
+  MIN_FILE_SIZE_STRING,
+} from '@whitewater-guide/commons';
 import FontIcon from 'material-ui/FontIcon';
 import React from 'react';
 import Dz from 'react-dropzone';
@@ -24,6 +30,9 @@ const styles: Styles = {
   icon: {
     fontSize: 64,
   },
+  error: {
+    color: 'red',
+  },
 };
 
 interface Props {
@@ -31,7 +40,13 @@ interface Props {
   onAdd?: (kind: MediaKind, file?: FileWithPreview) => void;
 }
 
-class Dropzone extends React.PureComponent<Props> {
+interface State {
+  error: string | null;
+}
+
+class Dropzone extends React.PureComponent<Props, State> {
+  readonly state: State = { error: null };
+
   onClick = () => {
     if (this.props.onAdd) {
       this.props.onAdd(this.props.kind);
@@ -40,12 +55,21 @@ class Dropzone extends React.PureComponent<Props> {
 
   onDrop = (acceptedFiles: File[]) => {
     if (this.props.onAdd) {
-      this.props.onAdd(this.props.kind, withPreview(acceptedFiles[0]));
+      const file = acceptedFiles[0];
+      if (file.size <= MIN_FILE_SIZE || file.size >= MAX_FILE_SIZE) {
+        this.setState({
+          error: `Min ${MIN_FILE_SIZE_STRING}, max ${MAX_FILE_SIZE_STRING}`,
+        });
+      } else {
+        this.setState({ error: null });
+        this.props.onAdd(this.props.kind, withPreview(file));
+      }
     }
   };
 
   render() {
     const { kind } = this.props;
+    const { error } = this.state;
     if (kind === MediaKind.photo) {
       return (
         <Dz onDrop={this.onDrop} multiple={false}>
@@ -61,6 +85,7 @@ class Dropzone extends React.PureComponent<Props> {
                   add
                 </FontIcon>
                 <span>or drop</span> image here
+                {!!error && <span style={styles.error}>{error}</span>}
               </div>
             );
           }}
@@ -77,6 +102,7 @@ class Dropzone extends React.PureComponent<Props> {
             add
           </FontIcon>
           <span>add video</span>
+          {!!error && <span style={styles.error}>{error}</span>}
         </div>
       );
     }