+++ services/web/src/utils/index.ts
@@ -3,3 +3,4 @@ export * from './files';
 export * from './paths';
 export * from './emitter';
 export * from './getImageSize';
+export * from './SentryRouterBreadcrumbs';