+++ packages/commons/CHANGELOG.md
@@ -1,3 +1,21 @@
+## [0.0.142](https://github.com/doomsower/whitewater/compare/@whitewater-guide/commons@0.0.141...@whitewater-guide/commons@0.0.142) (2019-06-26)
+
+### Features
+
+- **commons:** lower max image size to 3Mb ([9788717](https://github.com/doomsower/whitewater/commit/9788717))
+
+## [0.0.141](https://github.com/doomsower/whitewater/compare/@whitewater-guide/commons@0.0.139...@whitewater-guide/commons@0.0.141) (2019-06-26)
+
+### Features
+
+- **commons:** add min/max file size constants ([0f8e996](https://github.com/doomsower/whitewater/commit/0f8e996))
+
+## [0.0.140](https://github.com/doomsower/whitewater/compare/@whitewater-guide/commons@0.0.139...@whitewater-guide/commons@0.0.140) (2019-06-26)
+
+### Features
+
+- **commons:** add min/max file size constants ([8406458](https://github.com/doomsower/whitewater/commit/8406458))
+
 ## [0.0.139](https://github.com/doomsower/whitewater/compare/@whitewater-guide/commons@0.0.137...@whitewater-guide/commons@0.0.139) (2019-06-25)
 
 ### Bug Fixes