+++ services/backend/CHANGELOG.md
@@ -1,3 +1,11 @@
+## [0.0.362](https://github.com/doomsower/whitewater/compare/@whitewater-guide/backend@0.0.361...@whitewater-guide/backend@0.0.362) (2019-06-26)
+
+## [0.0.361](https://github.com/doomsower/whitewater/compare/@whitewater-guide/backend@0.0.360...@whitewater-guide/backend@0.0.361) (2019-06-26)
+
+### Features
+
+- **commons:** add min/max file size constants ([0f8e996](https://github.com/doomsower/whitewater/commit/0f8e996))
+
 ## [0.0.360](https://github.com/doomsower/whitewater/compare/@whitewater-guide/backend@0.0.359...@whitewater-guide/backend@0.0.360) (2019-06-25)
 
 ### Features