+++ services/boompromo/CHANGELOG.md
@@ -1,3 +1,15 @@
+## [0.0.167](https://github.com/doomsower/whitewater/compare/@whitewater-guide/boompromo@0.0.166...@whitewater-guide/boompromo@0.0.167) (2019-06-26)
+
+## [0.0.166](https://github.com/doomsower/whitewater/compare/@whitewater-guide/boompromo@0.0.165...@whitewater-guide/boompromo@0.0.166) (2019-06-26)
+
+### Features
+
+- **commons:** add min/max file size constants ([0f8e996](https://github.com/doomsower/whitewater/commit/0f8e996))
+
+## [0.0.165](https://github.com/doomsower/whitewater/compare/@whitewater-guide/boompromo@0.0.164...@whitewater-guide/boompromo@0.0.165) (2019-06-26)
+
+## [0.0.164](https://github.com/doomsower/whitewater/compare/@whitewater-guide/boompromo@0.0.163...@whitewater-guide/boompromo@0.0.164) (2019-06-26)
+
 ## [0.0.163](https://github.com/doomsower/whitewater/compare/@whitewater-guide/boompromo@0.0.162...@whitewater-guide/boompromo@0.0.163) (2019-06-25)
 
 ### Bug Fixes